soa-users-app/requirements.txt
fastapi[all]
uvicorn
alembic
pydantic
python-dotenv
python-keycloak
containers
dependency_injector